+++ integrations/sync/server/server.ts
@@ -7,45 +7,54 @@ type Response = {
   steamId: string | null;
   discordId: string | null;
   permissions: string[];
-  // todo: add types
-  unit: any;
 } | null;
 
+function prependSnailyCAD(text: string) {
+  return `^8^*[SnailyCAD]:^7^r ${text}`;
+}
+
+/**
+ * authentication flow
+ */
 RegisterCommand(
-  "whoami",
+  "sn-whoami",
   async (source: number) => {
     CancelEvent();
 
     const identifiers = getPlayerIds(source, "object");
-    const response = await cadRequest("/dispatch/player", "POST", identifiers).catch((error) => {
-      console.error(error);
-      return null;
-    });
+    const userId = identifiers.license;
+    const apiToken = GetResourceKvpString(`snailycad:${userId}:token`);
 
-    console.log("identifiers", identifiers);
+    const response = await cadRequest("/user", "POST", { userApiToken: apiToken }).catch(
+      (error) => {
+        console.error(error);
+        return null;
+      },
+    );
 
     const data = (await response?.body.json()) as Response;
 
-    console.log(data);
-
-    if (!data) {
+    if (!data?.id) {
+      emitNet("chat:addMessage", source, {
+        args: [prependSnailyCAD("Please make sure you're authenticated. Use: ^5/sn-auth^7.")],
+      });
       // todo: send client event that user doesn't exist
       return;
     }
 
     emitNet("chat:addMessage", source, {
-      args: [`Your username is ${data.username} and user ID is ${data.id}`],
+      args: [
+        prependSnailyCAD(
+          `Your SnailyCAD username is ^5${data.username} ^7and user ID is ^5${data.id}^7.`,
+        ),
+      ],
     });
   },
   false,
 );
 
-/**
- * authentication flow
- */
-
 RegisterCommand(
-  "authenticate",
+  "sn-auth",
   (source: number) => {
     CancelEvent();
 
@@ -55,9 +64,12 @@ RegisterCommand(
   false,
 );
 
-onNet("sna-sync:request-user-save", async (userData: unknown) => {
-  console.log("user", userData);
+onNet("sna-sync:request-user-save", async (userData: { token: string }) => {
+  const identifiers = getPlayerIds(source, "object");
+  if (!identifiers.license) {
+    console.error("no license found");
+    return;
+  }
 
-  // todo: validate user input
-  // todo: save in SQLite database
+  SetResourceKvp(`snailycad:${identifiers.license}:token`, userData.token);
 });